requirements.txt
git+https://github.com/openai/whisper.git
moviepy
ffmpeg-python
spacy
numpy
matplotlib
pydub